requirements.txt
ddgs>=9.9.1
python-dotenv>=0.9.9
duckduckgo-search>=8.1.1
fastapi>=0.121.2
langchain>=1.0.7
langchain-community>=0.4.1
langchain-core>=1.0.5
langchain-google-genai>=3.1.0
langgraph>=1.0.3
ollama>=0.6.1
pydantic>=2.12.4
requests>=2.32.5
streamlit>=1.51.0
uvicorn>=0.38.0
#vectorstores
langchain-pinecone
#embeddings models
langchain-cohere